backend/requirements.txt
fastapi
uvicorn
websockets
requests
pydantic
openai
murf
whisper
sounddevice
numpy
httpx
python-dotenv 
python-jose